requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
newsapi-python==0.2.7
httpx==0.25.2
email-validator==2.1.0.post1
pydantic==2.5.2
bcrypt==4.0.1
groq==0.3.0 